model folder renamed to dto
