+++ src/dto/bullet.ts
@@ -1,4 +1,4 @@
-import Vector from "../core/math/vector.js";
+import Vector from "../core/math/vector";
 import TransformModel from "./transform.js";
 
 export default interface BulletModel extends TransformModel {
